Enforce unique column labels.
An exception is now raised if column labels are not unique.

This allows for more obvious behaviour for the Tabulo::Row#to_h method.

This commit involves some refactoring, with columns now internally
stored in a Hash rather than an Array, and labels no longer
stored on the Column object itself.