Expose Table#shrink_to as a public method
This allows the table to be shrunk to a certain width, or to the width
of the terminal, while excluding certain columns from being resized in
the process, if desired, using the 'except' param.

Combined with the '#autosize_columns' method, it is now possible
to have fine-grained control over the resizing actions performed
by pack, by excluding specific columns from certain of those actions.